wip: move authz cli tests to e2e tests
