+++ tests/e2e/authz/query.go
@@ -1,4 +1,4 @@
-package testutil
+package authz
 
 import (
 	"fmt"
